+++ templates/part.settings.php
@@ -25,6 +25,7 @@
                 <button id="importDatasetButton" style="padding: 0 25px;">
                     <?php p($l->t('Import report')); ?>
                 </button>
+                <input type="file" id="importFile" accept="text/plain" hidden>
             </li>
             <li class="analytics-settings-item icon-external">
                 <a href="https://github.com/rello/analytics/wiki/donate" target="_blank" style="padding: 0 25px;">