+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - error handling when data source type is not available anymore
 - automated data deletion not working #339
 - Cannot change color of data series #338
+- chart legend not scrolling
 
 ## 4.7.2 - 2023-02-13
 ### Fixed